make randomize take 64 bit count to allow larger images to be tested BUG=none TESTED=untested
Review URL: https://webrtc-codereview.appspot.com/53679004

git-svn-id: http://libyuv.googlecode.com/svn/trunk@1429 16f28f9a-4ce2-e073-06de-1de4eb20be90

+++ unit_test/unit_test.h
@@ -58,8 +58,8 @@ static inline double get_time() {
 }
 #endif
 
-static inline void MemRandomize(uint8* dst, int len) {
-  int i;
+static inline void MemRandomize(uint8* dst, int64 len) {
+  int64 i;
   for (i = 0; i < len - 1; i += 2) {
     *reinterpret_cast<uint16*>(dst) = random();
     dst += 2;
